reorganize components to be more consistent; move HeroSection to a functional component
