+++ packages/react-server-website/components/homepage-body.less
@@ -1,4 +1,4 @@
-@import '../../middleware/Theme/constants';
+@import '../middleware/Theme/constants';
 
 section.HeroSection {
 	display: table;
