+++ packages/react-server-website/components/page-title.less
@@ -1,6 +1,6 @@
 @import "../middleware/Theme/constants";
 
-.DocTitle {
+.PageTitle {
 	padding: @small-section-pad 0;
 	.light-gray-to-black-background-color-gradient();
