+++ packages/react-server-website/components/source-body.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import './Docco.less';
+import './source-body.less';
 
-export default function Docco ({text}) {
+export default function SourceBody ({text}) {
 	return (
 		<div dangerouslySetInnerHTML={{ __html: text}}></div>
 	);
