+++ packages/react-server-website/components/source-body.less
@@ -1,5 +1,7 @@
 @import "../middleware/Theme/constants";
 
+// Styles are the gutted remains of Docco's default css
+
 /*--------------------- Layout ----------------------------*/
 #container { min-height: 100%; }
