+++ packages/react-server-website/middleware/PageHeader.js
@@ -1,22 +1,14 @@
 import React from 'react';
 import Header from '../components/Header';
 
-export default class PageHeader extends React.Component {
-
+export default class PageHeader {
 	getElements(next) {
 		var elements = [];
-
-		// skip adding header for fragments
-		if (this.getConfig('isFragment')) {
-			return next();
-		}
 
 		elements.push(
 			<Header />
 		)
 
-
 		return elements.concat(next());
 	}
-
 }
